// src/app/types/enums.ts
export enum Locale {
  rus = 'ru-RU',
  eng = 'en-EN',
}

export enum MaxSize {
  image = 100 * 1024,
}

export enum Keys {
  date = '61d2fe6b-13a6-4825-9bff-5c56d249aecb',
  days = 'e8bbcb08-8227-4f83-9e04-0101e6690028',
}

export enum QuestionType {
  text = 'TEXT',
  rating = 'RATING',
  radio = 'RADIO',
  checkbox = 'CHECKBOX',
}